(refactor) Extract `Creature` stamp to separate file
